+++ LibreriaSimulacion/Probadores/ProbadorExponencial.cs
@@ -12,21 +12,21 @@ namespace Numeros_aleatorios.LibreriaSimulacion.Probadores
     class ProbadorExponencial : IProbador
     {
         private DataTable numeros;
-        private float[] inicioIntervalos;
+        private double[] inicioIntervalos;
 
-        private float[] finIntervalos;
+        private double[] finIntervalos;
         private int[] frecuenciasObservadas;
-        private float[] frecuenciasEsperadas;
-        private float[] probabilidades;
+        private double[] frecuenciasEsperadas;
+        private double[] probabilidades;
         private DataTable resultado;
         private Truncador truncador;
-        private float valorCritico;
+        private double valorCritico;
         private double lambda;
         private double media;
 
         public ProbadorExponencial(Truncador truncador, DataTable numeros,
                                    double media, double lambda,
-                                   int cantidadIntervalos, float[] inicioIntervalos, float[] finIntervalos, int[] frecuenciasObservadas)
+                                   int cantidadIntervalos, double[] inicioIntervalos, double[] finIntervalos, int[] frecuenciasObservadas)
         {
             this.numeros = numeros;
             this.truncador = truncador;
@@ -53,7 +53,7 @@ namespace Numeros_aleatorios.LibreriaSimulacion.Probadores
             return compararEstadisticoConAcumulado();
         }
 
-        public float getValorCritico()
+        public double getValorCritico()
         {
             return valorCritico;
         }
@@ -67,9 +67,9 @@ namespace Numeros_aleatorios.LibreriaSimulacion.Probadores
             return resultado;
         }
 
-        public float obtenerTotalAcumuladoEstadisticoPrueba()
+        public double obtenerTotalAcumuladoEstadisticoPrueba()
         {
-            return float.Parse(resultado.Rows[resultado.Rows.Count - 2][6].ToString());
+            return double.Parse(resultado.Rows[resultado.Rows.Count - 2][6].ToString());
         }
 
         public void probar()
@@ -85,9 +85,9 @@ namespace Numeros_aleatorios.LibreriaSimulacion.Probadores
             this.resultado = new DataTable();
             crearTabla(resultado);
             DataRow row;
-            float marcaClase;
+            double marcaClase;
             double probabilidad;
-            float cantidadNumeros = numeros.Rows.Count;
+            double cantidadNumeros = numeros.Rows.Count;
             double frecuenciaEsperada;
 
             for (int i = 0; i < inicioIntervalos.Length; i++)
@@ -111,7 +111,7 @@ namespace Numeros_aleatorios.LibreriaSimulacion.Probadores
 
             }
         }
-        private float obtenerValorCritico(int gradosLibertad)
+        private double obtenerValorCritico(int gradosLibertad)
         {
             return ValorCriticoChi2.obtenerValorCritico(gradosLibertad);
         }
@@ -132,37 +132,37 @@ namespace Numeros_aleatorios.LibreriaSimulacion.Probadores
             return this.frecuenciasObservadas;
         }
 
-        public float[] getInicioIntervalos()
+        public double[] getInicioIntervalos()
         {
             return this.inicioIntervalos;
         }
 
-        public float[] getFinIntervalos()
+        public double[] getFinIntervalos()
         {
             return this.finIntervalos;
         }
 
         private void reestructurarTabla()
         {
-            float esperadaTablaVieja;
+            double esperadaTablaVieja;
             double esperadaAcumulada = 0;
             int observadaAcumulada = 0;
-            float probabilidadTablaVieja;
+            double probabilidadTablaVieja;
             double probabilidadAcumulada = 0;
-            List<float> nuevoInicioIntervalos = new List<float>();
-            List<float> nuevoFinIntervalos = new List<float>();
+            List<double> nuevoInicioIntervalos = new List<double>();
+            List<double> nuevoFinIntervalos = new List<double>();
             List<int> nuevaFrecuenciaObservada = new List<int>();
-            List<float> nuevaFrecuenciaEsperada = new List<float>();
-            List<float> nuevaProbabilidad = new List<float>();
-            float nuevoInicioIntervalo = 0;
-            float nuevoFinIntervalo = 0;
+            List<double> nuevaFrecuenciaEsperada = new List<double>();
+            List<double> nuevaProbabilidad = new List<double>();
+            double nuevoInicioIntervalo = 0;
+            double nuevoFinIntervalo = 0;
 
             for (int i = 0; i < inicioIntervalos.Length; i++)
             {
                 if (esperadaAcumulada == 0) { nuevoInicioIntervalo = inicioIntervalos[i]; }
-                esperadaTablaVieja = float.Parse(resultado.Rows[i][4].ToString());
+                esperadaTablaVieja = double.Parse(resultado.Rows[i][4].ToString());
                 esperadaAcumulada += esperadaTablaVieja;
-                probabilidadTablaVieja = float.Parse(resultado.Rows[i][3].ToString());
+                probabilidadTablaVieja = double.Parse(resultado.Rows[i][3].ToString());
                 probabilidadAcumulada += probabilidadTablaVieja;
                 observadaAcumulada += frecuenciasObservadas[i];
 
@@ -202,8 +202,8 @@ namespace Numeros_aleatorios.LibreriaSimulacion.Probadores
             DataRow row;
             double estadisticoPrueba;
             double estadisticoPruebaAcumuladoAnterior = 0;
-            float marcaClase;
-            float frecuenciaEsperada;
+            double marcaClase;
+            double frecuenciaEsperada;
 
             for (int i = 0; i < inicioIntervalos.Length; i++)
             {
